refined and split into modules to be more oop. added user input validators
// TODO: allow ESC out the flow

+++ Week03/ProjectWithSqlite/Entities/Entities.cs
@@ -4,7 +4,7 @@ namespace Entities {
   public class Product {
     public int id { get; set; }
     public string? name { get; set; }
-    public int price_in_gbp { get; set; }
+    public float price_in_gbp { get; set; }
     public string? description { get; set; }
     public bool availability { get; set; }
     public int stock_level { get; set; }
@@ -12,4 +12,3 @@ namespace Entities {
     public int promo_rate_as_percentage { get; set; }
   }
 }
-
